動画フィルターのログのviewerを実装
